Refactor login page into reusable LoginForm component
